Line-wrap normal document text and format headings
Add line-wrapping to Notes (normal document text). Centering is based
on the console width detected.
Add an example document for line wrapping.
Add recognition of nicknames in command line processing.
Update nicknames example to exercise hidden blocks from the command line.
Headings are now centered, the text is case-folded and the color
(foreground and background) is according to the level.
The optional prompt to exit after execution is now more frequent.